requirements.txt
numpy == 1.18.1
torch == 1.8.1   -f  https://download.pytorch.org/whl/torch_stable.html
torchvision == 0.4.2 
scipy == 1.4.1 
Pillow == 7.0.0
matplotlib == 3.1.3
opencv-python == 4.5.1.48